# [Optimization] Re-use converter functions across interpreters