## Description

Creating new interpreters/sub-interpreters can be optimized by reusing the converter functions across interpreters.

I assumed it is safe to re-use them since: 
- They are stateless functions
- Nested variables are read-only

### Performance

```
name                                   old time/op    new time/op    delta
NewInterpreter/new_interpreter-12        20.6µs ±16%     7.7µs ±16%  -62.84%  (p=0.008 n=5+5)
NewInterpreter/new_sub-interpreter-12    20.8µs ±23%     7.7µs ±12%  -62.88%  (p=0.008 n=5+5)

name                                   old alloc/op   new alloc/op   delta
NewInterpreter/new_interpreter-12        15.9kB ± 0%     5.3kB ± 0%  -66.62%  (p=0.008 n=5+5)
NewInterpreter/new_sub-interpreter-12    16.3kB ± 0%     5.6kB ± 0%  -65.71%  (p=0.008 n=5+5)

name                                   old allocs/op  new allocs/op  delta
NewInterpreter/new_interpreter-12           274 ± 0%        66 ± 0%  -75.91%  (p=0.008 n=5+5)
NewInterpreter/new_sub-interpreter-12       292 ± 0%        83 ± 0%  -71.58%  (p=0.008 n=5+5)

```

______

<!-- Complete: -->

- [x] Targeted PR against `master` branch
- [x] Linked to Github issue with discussion and accepted design OR link to spec that describes this work
- [x] Code follows the [standards mentioned here](https://github.com/onflow/cadence/blob/master/CONTRIBUTING.md#styleguides)
- [x] Updated relevant documentation 
- [x] Re-reviewed `Files changed` in the Github PR explorer
- [x] Added appropriate labels